+++ tests/clients/test_instance.py
@@ -180,16 +180,33 @@ class TestInstanceClient:
             self._delete_user(client, user_1)
             self._delete_user(client, user_2)
 
-    def test_update_instance_user(self, client: MySQLInstanceClient):
-        """Test the updating of an instance user."""
+    def test_update_instance_user_password(self, client: MySQLInstanceClient):
+        """Test the updating of an instance user password."""
         instance_user = User("instance_user_update", "%")
 
         try:
-            client.create_instance_user(instance_user, "password")
-            client.update_instance_user(instance_user, "password_new")
+            client.create_instance_user(instance_user, password="password")
+            client.update_instance_user(instance_user, password="password_new")
 
             executor = build_local_executor(instance_user.username, "password_new")
             executor.check_connection()
+        finally:
+            self._delete_user(client, instance_user)
+
+    def test_update_instance_user_attributes(self, client: MySQLInstanceClient):
+        """Test the updating of an instance user attributes."""
+        old_attrs = {"key": "val_1"}
+        new_attrs = {"key": "val_2"}
+
+        instance_user = User("instance_user_update", "%", old_attrs)
+
+        try:
+            client.create_instance_user(instance_user, password="password")
+            client.update_instance_user(instance_user, password="password", attrs=new_attrs)
+
+            users = client.search_instance_users(instance_user.username)
+            assert len(users) > 0
+            assert users[0].attributes == new_attrs
         finally:
             self._delete_user(client, instance_user)
 
@@ -253,6 +270,21 @@ class TestInstanceClient:
         """Test the reloading of instance TLS certificates."""
         pass
 
+    def test_search_instance_replication_members(self, client: MySQLInstanceClient):
+        """Test the searching of instance replication members."""
+        members = client.search_instance_replication_members()
+        assert len(members) == 1
+
+        primary_members = client.search_instance_replication_members([InstanceRole.PRIMARY], [])
+        replica_members = client.search_instance_replication_members([InstanceRole.SECONDARY], [])
+        assert len(primary_members) == 1
+        assert len(replica_members) == 0
+
+        online_members = client.search_instance_replication_members([], [InstanceState.ONLINE])
+        offline_members = client.search_instance_replication_members([], [InstanceState.OFFLINE])
+        assert len(online_members) == 1
+        assert len(offline_members) == 0
+
     def test_search_instance_connections(self, client: MySQLInstanceClient):
         """Test the searching of instance connections given a name-pattern."""
         query = "DO SLEEP(10)"
@@ -261,8 +293,8 @@ class TestInstanceClient:
             process_ids = self._get_processes(client, query)
             assert len(process_ids) == 1
 
-            assert process_ids[0] in client.search_instance_connections("%")
-            assert process_ids[0] not in client.search_instance_connections("process_search")
+            assert process_ids[0] in client.search_instance_connection_processes("%")
+            assert process_ids[0] not in client.search_instance_connection_processes("search")
 
     def test_search_instance_databases(self, client: MySQLInstanceClient):
         """Test the searching of instance databases given a name-pattern."""
@@ -332,6 +364,6 @@ class TestInstanceClient:
             process_ids = self._get_processes(client, query)
             assert len(process_ids) == 1
 
-            assert process_ids[0] in client.search_instance_connections("%")
+            assert process_ids[0] in client.search_instance_connection_processes("%")
             client.stop_instance_processes(process_ids)
-            assert process_ids[0] not in client.search_instance_connections("%")
+            assert process_ids[0] not in client.search_instance_connection_processes("%")