+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog][docs-changelog], and the version adher
 
 
 ## Unreleased
+### Added
+- Search replication members method to InstanceClient class.
+- Ability to update user attrs by the InstanceClient class.
 ### Fixed
 - Rename InstanceStatus to InstanceState.
 - Crash when querying instance role and instance state.